requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
category-encoders==2.3.0
certifi==2021.5.30
cffi==1.15.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==4.0
djangorestframework==3.13.1
entrypoints==0.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy @ file:///Users/ktietz/demo/mc3/conda-bld/numpy_1629115584384/work
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==1.11.3
scikit-learn @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/scikit-learn_1635339011232/work
scipy @ file:///Users/ktietz/demo/mc3/conda-bld/scipy_1629907664009/work
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
statsmodels==0.10.2
terminado==0.12.1
testpath==0.5.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.6.0
